pyproject.toml
[project]
name = "rogue-scroll"
dynamic = ["version"]
description = "Generate scroll titles from the game rogue"
readme = "README.md"
license = "CC-BY-ND-4.0"
authors = [
    { name = "Jeffrey Goldberg", email = "jeffrey@goldmark.org" },
    { name = "1Password Inc" }
]
maintainers = [
    { name = "Jeffrey Goldberg", email = "jeffrey@goldmark.org" }
]
requires-python = ">=3.10"
dependencies = [
    "pandas-stubs>=2.2.3.241126",
    "seaborn>=0.13.2",
]

classifiers = [
    "Development Status :: 2 - Pre-Alpha",
]

[project.urls]
Repository = "https://github.com/jpgoldberg/rogue-scroll.git"
Issues = "https://github.com/jpgoldberg/rogue-scroll/issues"


[project.scripts]
rogue-scroll = "rogue_scroll:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "src/rogue_scroll/__about__.py"

[tool.ruff]
line-length = 80

[tool.mypy]
strict = true
